+++ src/z80asm/t/op-ASMSIZE-ASMTAIL.t
@@ -14,8 +14,8 @@ use warnings;
 use Test::More;
 require 't/test_utils.pl';
 
-my @testfiles = qw( testa.asm testa.lis testa.sym testa.obj testa.map testa.bin
-					testb.asm testb.lis testb.sym testb.obj
+my @testfiles = qw( testa.asm testa.lis testa.sym testa.o testa.map testa.bin
+					testb.asm testb.lis testb.sym testb.o
 				);
 
 my $asm = "